% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.admissions.r
\name{add.admissions}
\alias{add.admissions}
\title{Add an "admissions" column}
\usage{
add.admissions(X)
}
\arguments{
\item{X}{a data frame}
}
\value{
An identical data frame with an "admissions" column.
}
\description{
Add an "admissions" column
}
\details{
Add a column with a the number of admissions, 
computed as \code{diff(X$hosp) + diff(X$rad) + diff(X$dc)}.
Special care is given to the case of missing days.
}
