% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.r
\name{agg}
\alias{agg}
\title{Data aggregation}
\usage{
agg(x, cols, by, FUN = sum, ...)
}
\arguments{
\item{x}{a data frame}

\item{cols}{the columns to aggregate}

\item{by}{grouping elements}

\item{FUN}{function used to aggregate the data}

\item{...}{further parameters passed to `aggregate` or to `FUN`}
}
\value{
a data frame, with the data in `x[,cols]\` aggregated according 
to the levels of the factors in `x[,by]`.
}
\description{
Data aggregation
}
\details{
This function is a convenient wrapper for \code{\link{aggregate}}.
It is implemented as `aggregate( x[,cols], x[, by], FUN, ...)`.
}
\examples{
agg(departements, cols = c("superficie", "population"), by = "région")

}
