% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.dept.r
\name{as.dept}
\alias{as.dept}
\title{Creates a object of class \code{dept}}
\usage{
as.dept(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
an object of class \code{dept}
}
\description{
Creates a object of class \code{dept}
}
\details{
The class \code{dept} allow to handle french départements. 
Methods have been implemented (in particular for \code{xtfrm}) 
to ensure that départements \code{"2A"} and \code{"2B"} are inserted between
\code{"19"} and \code{"21"} when sorting an object of class \code{dept}.
}
