% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{departements}
\alias{departements}
\title{French départements}
\format{
data frame
}
\usage{
data(departements)
}
\description{
French départements
}
\details{
This dataset contains data about the French départements. The colmun names
are self-documenting.
}
\keyword{departements}
