% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.spf.r
\name{read.spf}
\alias{read.spf}
\title{Read Santé Publique France data}
\usage{
read.spf(file = c("-", "nouveaux", "classe-age", "etablissements"))
}
\arguments{
\item{file}{which file to read}
}
\value{
A data frame
}
\description{
Read Santé Publique France data
}
\details{
This function reads one the files 
described at \url{https://www.data.gouv.fr/fr/datasets/donnees-hospitalieres-relatives-a-lepidemie-de-covid-19/}
and do some cleaning.
}
\examples{
covid <- read.spf("nouveaux")
covid <- agg(covid, 3:6, 2)
covid$rolling_hosp <- rolling.mean(covid$incid_hosp, 7, "right")
covid$rolling_rea  <- rolling.mean(covid$incid_rea,  7, "right")
covid$rolling_dc   <- rolling.mean(covid$incid_dc,   7, "right")
covid$rolling_rad  <- rolling.mean(covid$incid_rad,  7, "right")


plot(covid$jour, covid$incid_dc, type = "h")
lines(covid$jour, covid$rolling_dc, col = "red", lwd = 2)

}
