% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{regions}
\alias{regions}
\title{French régions}
\format{
data frame
}
\usage{
data(regions)
}
\description{
French régions
}
\details{
This dataset contains the numerical codes for the French 
administrative regions.
}
\keyword{regions}
