% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling.mean.r
\name{rolling.mean}
\alias{rolling.mean}
\title{Rolling mean}
\usage{
rolling.mean(x, n = 7, align = c("center", "left", "right"))
}
\arguments{
\item{x}{a numeric vector}

\item{n}{an integer}

\item{...}{extra parameters for \code{filter}}
}
\value{
The rolling mean of \code{x} based on \code{n} points. 
This is just a wrapper for \code{as.vector(filter(x, rep(n,1/n), ...))}.
}
\description{
Rolling mean
}
