% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.by.depts.r
\name{summarize.by.depts}
\alias{summarize.by.depts}
\title{Summarizing covid data by region}
\usage{
summarize.by.depts(x, departements, n = 7)
}
\arguments{
\item{x}{a data frame}

\item{departements}{a vector of departments, defining a region}

\item{n}{length of rolling average}
}
\value{
a data frame
}
\description{
Summarizing covid data by region
}
\details{
This function summarizes data obtained from \code{read.spf("nouveaux")},
aggregating data from the specified departements and adding a n-last-days rolling 
average. Default is \code{n = 7}.
}
\examples{
Data <- read.spf("nouveaux")
# summarizes for department = 75
Data.75 <- summarize.by.depts(Data, 75)
tail(Data.75, 10)
# for Ile-de-France
idf <- departements$numéro[ departements$code_région == 11 ] 
Data.idf <- summarize.by.depts(Data, idf)
tail(Data.idf, 10)
# plot for last 100 days
X <- tail(Data.idf, 100)
plot(X$jour, X$incid_hosp, log = "y", type = "o")
lines(X$jour, X$rolling_hosp, lwd = 2, col = "red")

}
