% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.spf.r
\name{read.spf}
\alias{read.spf}
\title{Read Santé Publique France data}
\usage{
read.spf(file = c("-", "nouveaux", "classe-age", "etablissements"))
}
\arguments{
\item{file}{which file to read}
}
\value{
A data frame
}
\description{
Read Santé Publique France data
}
\details{
This function reads one the files 
described at \url{https://www.data.gouv.fr/fr/datasets/donnees-hospitalieres-relatives-a-lepidemie-de-covid-19/}
and do some cleaning.
}
\examples{
cov <- read.spf("nouveaux")
cov <- agg(cov, 3:6, 2)
plot(cov$jour, cov$incid_rea, log = "y", type = "o")
lines(cov$jour, rolling.mean(cov$incid_rea), col = "red", lwd = 2)

}
